requirements.txt
# -f git+https://github.com/facebookresearch/detectron2.git

accelerate==0.15.0
camelot-py[cv]==0.10.1
CurrencyConverter==0.17.1
datasets==2.4.0
django==3.2.15
djangorestframework==3.13.1
fairscale==0.4.6
fvcore==0.1.5.post20220512
gensim==4.2.0
language-tool-python==2.7.1
nltk==3.7
omegaconf==2.2.2
opencv-python==4.6.0.66
paddlepaddle==2.3.2
paddlenlp==2.4.1
pandas==1.3.5
pdf2image==1.16.0
Pint==0.18
psycopg2-binary==2.8.6
pymupdf==1.20.0
pytz==2022.6
pyyaml
pytesseract==0.3.9
sentence-transformers==2.2.2
seqeval==1.2.2
timm==0.6.5
torch==1.8.0
torchvision==0.13.1
transformers==4.25.1
